Refactor project inheritance
