+++ src/Fonts.ruleset
@@ -4,6 +4,7 @@
   <Rules AnalyzerId="Microsoft.CodeAnalysis.CSharp.Features" RuleNamespace="Microsoft.CodeAnalysis.CSharp.Features">
     <Rule Id="IDE0010" Action="None" />
     <Rule Id="IDE0072" Action="None" />
+     <Rule Id="IDE0240" Action="None" />
   </Rules>
   <Rules AnalyzerId="StyleCop.Analyzers" RuleNamespace="StyleCop.Analyzers">
     <Rule Id="SA1011" Action="Info" />
